Restructuring the codebase! But for real, it's a real shitstorm in there. - Added context by making the init go through a function   and not implicitly happen via imports - Fixup all context and contextless function mixups - splitup the models in sensible different files - give the dump of view functions in views/__init__.py their own file - add all routes via blueprints, not half of them - move the slack notifications function and class to its own file,     no idea what it was doing in a views file in the first place.
