--- app/__init__.py
@@ -1,2 +1,0 @@
-# Bind  app.db to db
-from app.app import db as db
