+++ app/passenger_wsgi.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
 
-import sys
 import os
+import sys
+
+from app import create_app
 
 INTERP = os.path.expanduser("~/env/bin/python")
 if sys.executable != INTERP:
@@ -9,4 +11,8 @@ if sys.executable != INTERP:
 
 sys.path.append(os.getcwd())
 
-from haldis import app as application
+application = create_app()
+
+# For running on the server with passenger etc
+if __name__ == "__main__":
+    application.run(port=8000)
