+++ app/templates/errors/401.html
@@ -4,6 +4,6 @@
     <div class="jumbotron">
         <h1>Unauthorized</h1>
         <p>You're not authorized to look to this page!</p>
-        <p><a href="{{ url_for('home') }}">Go somewhere nice</a></p>
+        <p><a href="{{ url_for('general_bp.home') }}">Go somewhere nice</a></p>
     </div>
 {% endblock %}
