+++ app/templates/errors/404.html
@@ -4,6 +4,6 @@
     <div class="jumbotron">
         <h1>Page Not Found</h1>
         <p>What you were looking for is just not there.</p>
-        <p><a href="{{ url_for('home') }}">Go somewhere nice</a></p>
+        <p><a href="{{ url_for('general_bp.home') }}">Go somewhere nice</a></p>
     </div>
 {% endblock %}
