+++ app/templates/layout.html
@@ -2,12 +2,12 @@
 {% import "bootstrap/utils.html" as utils %}
 
 {% set navbar = [
-    ('home', 'Home'),
+    ('general_bp.home', 'Home'),
     ('order_bp.orders', 'Orders'),
-    ('locations', 'Locations'),
-    ('map', 'Map'),
-    ('about', 'About'),
-    ('stats', 'Stats'),
+    ('general_bp.locations', 'Locations'),
+    ('general_bp.map', 'Map'),
+    ('general_bp.about', 'About'),
+    ('stats_blueprint.stats', 'Stats'),
 ] -%}
 {% if not current_user.is_anonymous() and current_user.is_admin() -%}
     {% set navbar = navbar + [('admin.index', 'Admin')] -%}
@@ -43,7 +43,7 @@ Haldis - {{ active_page|capitalize }}
             <span class="icon-bar"></span>
             <span class="icon-bar"></span>
           </button>
-          <a class="navbar-brand" href="{{ url_for('home') }}">HALDIS</a>
+          <a class="navbar-brand" href="{{ url_for('general_bp.home') }}">HALDIS</a>
         </div>
         <div id="navbar" class="navbar-collapse collapse">
           <ul class="nav navbar-nav">
@@ -55,10 +55,10 @@ Haldis - {{ active_page|capitalize }}
           </ul>
           <ul class="nav navbar-nav navbar-right">
               {% if current_user.is_anonymous() %}
-                  <li><a href="{{ url_for('login') }}">Login</a></li>
+                  <li><a href="{{ url_for('auth_bp.login') }}">Login</a></li>
               {% else %}
-                  <li><a href="{{ url_for('profile') }}">{{ current_user.username }}</a></li>
-                  <li><a href="{{ url_for('logout') }}">Logout</a></li>
+                  <li><a href="{{ url_for('general_bp.profile') }}">{{ current_user.username }}</a></li>
+                  <li><a href="{{ url_for('auth_bp.logout') }}">Logout</a></li>
               {% endif %}
           </ul>
          </div><!--/.nav-collapse -->
