+++ app/templates/orders.html
@@ -26,7 +26,7 @@
             <h3>Create new order:</h3>
         <div class="row darker">
             <div class="col-sm-12">
-                <form method="post" action="{{ url_for('.order_create') }}">
+                <form method="post" action="{{ url_for('order_bp.order_create') }}">
                     {{ form.csrf_token }}
                     <div class="form-group select2 {{ 'has-errors' if form.courrier_id.errors else ''}}">
                         {{ form.courrier_id.label(class='control-label') }}<br>
