+++ app/views/order.py
@@ -1,17 +1,18 @@
 __author__ = 'feliciaan'
-import json
-from threading import Thread
-import requests
-from flask import url_for, render_template, abort, redirect, Blueprint, flash, session, request
-from flask_login import current_user, login_required
 import random
 from datetime import datetime
 
-from app import app, db
-from models import Order, OrderItem, User
-from forms import OrderItemForm, OrderForm, AnonOrderItemForm
+# from flask import current_app as app
+from flask import (Blueprint, abort, flash, redirect, render_template, request,
+                   session, url_for)
+from flask_login import current_user, login_required
+
+from forms import AnonOrderItemForm, OrderForm, OrderItemForm
+from models import Order, OrderItem, User, db
+from notification import post_order_to_webhook
 
 order_bp = Blueprint('order_bp', 'order')
+
 
 @order_bp.route('/')
 def orders(form=None):
@@ -35,7 +36,7 @@ def order_create():
         db.session.add(order)
         db.session.commit()
         post_order_to_webhook(order)
-        return redirect(url_for('.order', id=order.id))
+        return redirect(url_for('order_bp.order', id=order.id))
     return orders(form=orderForm)
 
 
@@ -48,13 +49,19 @@ def order(id, form=None):
         flash('Please login to see this order.', 'info')
         abort(401)
     if form is None:
-        form = AnonOrderItemForm() if current_user.is_anonymous() else OrderItemForm()
+        form = AnonOrderItemForm() if current_user.is_anonymous(
+        ) else OrderItemForm()
         form.populate(order.location)
     if order.stoptime and order.stoptime < datetime.now():
         form = None
     total_price = sum([o.product.price for o in order.items])
     debts = sum([o.product.price for o in order.items if not o.paid])
-    return render_template('order.html', order=order, form=form, total_price=total_price, debts=debts)
+    return render_template('order.html',
+                           order=order,
+                           form=form,
+                           total_price=total_price,
+                           debts=debts)
+
 
 @order_bp.route('/<id>/items')
 def items_showcase(id, form=None):
@@ -66,11 +73,13 @@ def items_showcase(id, form=None):
         abort(401)
     return render_template('order_items.html', order=order)
 
+
 @order_bp.route('/<id>/edit', methods=['GET', 'POST'])
 @login_required
 def order_edit(id):
     order = Order.query.filter(Order.id == id).first()
-    if current_user.id is not order.courrier_id and not current_user.is_admin():
+    if current_user.id is not order.courrier_id and not current_user.is_admin(
+    ):
         abort(401)
     if order is None:
         abort(404)
@@ -79,7 +88,7 @@ def order_edit(id):
     if orderForm.validate_on_submit():
         orderForm.populate_obj(order)
         db.session.commit()
-        return redirect(url_for('.order', id=order.id))
+        return redirect(url_for('order_bp.order', id=order.id))
     return render_template('order_edit.html', form=orderForm, order_id=id)
 
 
@@ -93,7 +102,8 @@ def order_item_create(id):
     if current_user.is_anonymous() and not current_order.public:
         flash('Please login to see this order.', 'info')
         abort(401)
-    form = AnonOrderItemForm() if current_user.is_anonymous() else OrderItemForm()
+    form = AnonOrderItemForm() if current_user.is_anonymous(
+    ) else OrderItemForm()
     form.populate(current_order.location)
     if form.validate_on_submit():
         item = OrderItem()
@@ -106,7 +116,7 @@ def order_item_create(id):
         db.session.add(item)
         db.session.commit()
         flash('Ordered %s' % (item.product.name), 'success')
-        return redirect(url_for('.order', id=id))
+        return redirect(url_for('order_bp.order', id=id))
     return order(id, form=form)
 
 
@@ -118,8 +128,9 @@ def item_paid(order_id, item_id):
     if item.order.courrier_id == id or current_user.admin:
         item.paid = True
         db.session.commit()
-        flash('Paid %s by %s' % (item.product.name, item.get_name()), 'success')
-        return redirect(url_for('.order', id=order_id))
+        flash('Paid %s by %s' % (item.product.name, item.get_name()),
+              'success')
+        return redirect(url_for('order_bp.order', id=order_id))
     abort(404)
 
 
@@ -129,9 +140,11 @@ def items_user_paid(order_id, user_name):
     user = User.query.filter(User.username == user_name).first()
     items = []
     if user:
-        items = OrderItem.query.filter((OrderItem.user_id == user.id) & (OrderItem.order_id == order_id))
+        items = OrderItem.query.filter((OrderItem.user_id == user.id)
+                                       & (OrderItem.order_id == order_id))
     else:
-        items = OrderItem.query.filter((OrderItem.name == user_name) & (OrderItem.order_id == order_id))
+        items = OrderItem.query.filter((OrderItem.name == user_name)
+                                       & (OrderItem.order_id == order_id))
     current_order = Order.query.filter(Order.id == order_id).first()
     for item in items:
         print(item)
@@ -139,8 +152,9 @@ def items_user_paid(order_id, user_name):
         for item in items:
             item.paid = True
         db.session.commit()
-        flash('Paid %d items for %s' % (items.count(), item.get_name()), 'success')
-        return redirect(url_for('.order', id=order_id))
+        flash('Paid %d items for %s' % (items.count(), item.get_name()),
+              'success')
+        return redirect(url_for('order_bp.order', id=order_id))
     abort(404)
 
 
@@ -156,7 +170,7 @@ def delete_item(order_id, item_id):
         db.session.delete(item)
         db.session.commit()
         flash('Deleted %s' % (product_name), 'success')
-        return redirect(url_for('.order', id=order_id))
+        return redirect(url_for('order_bp.order', id=order_id))
     abort(404)
 
 
@@ -172,7 +186,7 @@ def volunteer(id):
         flash("Thank you for volunteering!")
     else:
         flash("Volunteering not possible!")
-    return redirect(url_for('.order', id=id))
+    return redirect(url_for('order_bp.order', id=id))
 
 
 @order_bp.route('/<id>/close')
@@ -190,9 +204,7 @@ def close_order(id):
             if courrier is not None:
                 order.courrier_id = courrier.id
         db.session.commit()
-        return redirect(url_for('.order', id=id))
-
-app.register_blueprint(order_bp, url_prefix='/order')
+        return redirect(url_for('order_bp.order', id=id))
 
 
 def select_user(items):
@@ -216,52 +228,12 @@ def select_user(items):
 def get_orders(expression=None):
     orders = []
     if expression is None:
-        expression = ((datetime.now() > Order.starttime) & (Order.stoptime > datetime.now()) | (Order.stoptime == None))
+        expression = ((datetime.now() > Order.starttime) &
+                      (Order.stoptime > datetime.now()) |
+                      (Order.stoptime == None))
     if not current_user.is_anonymous():
         orders = Order.query.filter(expression).all()
     else:
-        orders = Order.query.filter((expression & (Order.public == True))).all()
+        orders = Order.query.filter(
+            (expression & (Order.public == True))).all()
     return orders
-
-
-def post_order_to_webhook(order_item):
-    message = ''
-    if order_item.courrier is not None:
-        message = '<!channel|@channel> {3} is going to {1}, order <{0}|here>! Deadline in {2} minutes!'.format(
-                url_for('.order', id=order_item.id, _external=True),
-                order_item.location.name,
-                remaining_minutes(order_item.stoptime),
-                order_item.courrier.username.title())
-    else:
-        message = '<!channel|@channel> New order for {}. Deadline in {} minutes. <{}|Open here.>'.format(
-                order_item.location.name,
-                remaining_minutes(order_item.stoptime),
-                url_for('.order', id=order_item.id, _external=True))
-    webhookthread = WebhookSenderThread(message)
-    webhookthread.start()
-
-
-class WebhookSenderThread(Thread):
-
-    def __init__(self, message):
-        super(WebhookSenderThread, self).__init__()
-        self.message = message
-
-    def run(self):
-        self.slack_webhook()
-
-    def slack_webhook(self):
-        js = json.dumps({'text': self.message})
-        url = app.config['SLACK_WEBHOOK']
-        if len(url) > 0:
-            requests.post(url, data=js)
-        else:
-            app.logger.info(str(js))
-
-
-def remaining_minutes(value):
-    delta = value - datetime.now()
-    if delta.total_seconds() < 0:
-        return "0"
-    minutes, _ = divmod(delta.total_seconds(), 60)
-    return "%02d" % minutes
